requirements.txt
openai-agents>=0.0.17
gradio>=6.3.0
python-dotenv>=1.0.0
sendgrid>=6.12.0
pydantic>=2.0.0
